Refactor handler to use abstracted common version
