+++ service-front/app/src/Actor/src/Handler/RequestActivationKey/AbstractRequestKeyHandler.php
@@ -1,0 +1,69 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Actor\Handler\RequestActivationKey;
+
+use Common\Handler\{AbstractHandler, CsrfGuardAware, Traits\CsrfGuard, Traits\Session as SessionTrait, UserAware};
+use Actor\Form\RequestActivationKey\RequestDateOfBirth;
+use Common\Handler\Traits\User;
+use Common\Service\Url\UrlValidityCheckService;
+use Laminas\Diactoros\Response\RedirectResponse;
+use Mezzio\Authentication\UserInterface;
+use Mezzio\Session\SessionInterface;
+use Psr\Http\Message\{ResponseInterface, ServerRequestInterface};
+use Mezzio\Authentication\AuthenticationInterface;
+use Mezzio\Helper\UrlHelper;
+use Mezzio\Template\TemplateRendererInterface;
+use Laminas\Diactoros\Response\HtmlResponse;
+
+/**
+ * Class RequestActivationKeyHandler
+ * @package Actor\Handler
+ * @codeCoverageIgnore
+ */
+abstract class AbstractRequestKeyHandler extends AbstractHandler implements UserAware, CsrfGuardAware
+{
+    use User;
+    use CsrfGuard;
+    use SessionTrait;
+
+    protected ?SessionInterface $session;
+    protected ?UserInterface $user;
+
+    public function __construct(
+        TemplateRendererInterface $renderer,
+        AuthenticationInterface $authenticator,
+        UrlHelper $urlHelper
+    ) {
+        parent::__construct($renderer, $urlHelper);
+
+        $this->setAuthenticator($authenticator);
+    }
+
+    /**
+     * @param ServerRequestInterface $request
+     * @return ResponseInterface
+     */
+    public function handle(ServerRequestInterface $request): ResponseInterface
+    {
+        $this->user = $this->getUser($request);
+        $this->session = $this->getSession($request, 'session');
+
+        switch ($request->getMethod()) {
+            case 'POST':
+                return $this->handlePost($request);
+            default:
+                return $this->handleGet($request);
+        }
+    }
+
+    abstract public function handleGet(ServerRequestInterface $request): ResponseInterface;
+
+    abstract public function handlePost(ServerRequestInterface $request): ResponseInterface;
+
+    protected function hasFutureAnswersInSession(): bool
+    {
+        return $this->session->get('postcode') != null;
+    }
+}
